New script to generate static maps and world files using mapserver (shp2img)
These maps are improved over the previous version b/c they set the map extent
in meters from the radar station instead of degrees lat/long (which are
inconsistent with the actual radar measurements).

The map background is transparent so they can be overlaid on top of radar
data (or vice versa).

The script is

convert/gen-static-maps.pl

The files generated by the script are in

label/site/static-maps